refactor(gateway): extract frontmatter parser to shared util
The agent loader was importing splitFrontmatter from the skill module and
duplicating parseYaml/parseScalar locally. The skill module is scheduled
for rename to "recipe" (progressive-disclosure playbooks, opencode pattern),
so the shared pieces move to a util both loaders can depend on without the
agent module reaching into the skill namespace.

No behavior change. agent/index.ts shrinks by 51 lines; same tests pass
(154/154).
